svr/requirement.txt
celery==4.0.2
Django==1.11.2
django-celery-beat==1.0.1
django-celery-results==1.0.1
django-redis-cache==1.7.1
djangorestframework==3.6.3
httplib2==0.10.3
MySQL-python==1.2.5
redis==2.10.5
uwsgi==2.0.15
supervisor==3.3.2
